requirements.txt
youtube-transcript-api
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-youtube
anthropic
google-genai
openai